// src/app/generated_services/model/updateNotificationDTO.ts
/**
 * BjjClutch Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { NotificationType } from './notificationType';


export interface UpdateNotificationDTO { 
    title: string;
    message: string;
    type: NotificationType;
    isActive?: boolean;
    expiresAt?: string | null;
    actionUrl?: string | null;
    metadata?: string | null;
}
export namespace UpdateNotificationDTO {
}